Parse Github Asset Based on Accessibility
For supporting private releases, the ReleaseAsset parsing function was
changed to use a different field in the JSON document, namely from
browser_download_url -> url. This change breaks the download of assets
for public Releases, as those still use the previous field.

You can see this in the documentation for the API:
https://docs.github.com/en/rest/reference/repos#releases

On a public release, the 'url' field on the asset doesn't return the
download location for the asset, so the download_url for the
ReleaseAsset struct isn't set correctly.

This change introduces an 'is_private' parameter on the from_asset
function set the field based on whether the release is private or not.
It's determined to be private based on whether an auth token has been
specified.