Add identifier for releases with multiple compatible assets
This change is relevant when you have multiple binaries as part of a
single release. In this case, the target is not enough to determine
which asset to choose. Thus this change allows an identifier to be
passed along from the builders, to select the correct one in such cases.

This adds argument `identifier: Option<&str>` to `Release::asset_for`,
which allows the caller to specify a string that must be part of the
asset name in addition to the target.

Additionally it adds a method to `ReleaseUpdate` which allows the
identifier to be specified in the `UpdateBuilder` structs.

One possible change, could be to make the `identifier` a closure, or
like the `Pattern` like the struct search functions in `std`, instead of
relying on only substrings.