fix: reqwest TLS backend
Currently, the rustls TLS backend from reqwest is always included and
used even when the `rustls` feature is _not_ enabled, via the
`use_rustls_tls` function call.

This is not only unexpected behaviour but also increases the size of
application binaries since both the default (native_tls) and rustls
backends are included.

When _only_ rustls is needed, use the `rustls` feature _and_ disable
the default features.